+++ CHANGELOG.md
@@ -2,11 +2,25 @@
 
 ## Unrealeased
 
+### Todo
+
+- Tutorial
+- Responsive Canvas
+- Demo mode
+- Multiple system management (load, save and display)
+
+### Added
+
+- Motionless bodies
+- Edit of body attributes
+- Body shadow
+- Demo mode (via URL parameter `demo`)
+
 ### Changed
 
 - Bump to new version of Konva.js (7.2.2)
 - Magnifing factor of vectors
-- Visibility factor of vectors
+- Visibility of vectors
 
 ## 0.1.0-alpha